# Conflicts:
#	pom.xml
#	src/main/java/com/example/repair/config/myMetaObjectHandler.java
#	src/main/java/com/example/repair/controller/maintainer/maintainerController.java
#	src/main/java/com/example/repair/entity/AdministratorAccount.java
#	src/main/java/com/example/repair/entity/MaintainerAccount.java
#	src/main/java/com/example/repair/entity/MaintenanceRecord.java
#	src/main/java/com/example/repair/entity/PreliminaryScheme.java
#	src/main/java/com/example/repair/entity/StudentAccount.java
#	src/main/java/com/example/repair/util/ResultCode.java
#	src/main/resources/application.yaml
#	src/test/java/com/example/Demo01ApplicationTests.java

+++ src/main/java/com/example/repair/entity/MaintainerAccount.java
@@ -31,7 +31,7 @@ public class MaintainerAccount implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "job_number", type = IdType.INPUT)
-    private Integer jobNumber;
+    private Long jobNumber;
 
     private String name;
