+++ src/main/java/com/example/repair/entity/MaintenanceRecord.java
@@ -30,10 +30,10 @@ public class MaintenanceRecord implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "FK_workorder_number", type = IdType.INPUT)
-    private Integer fkWorkorderNumber;
+    private Long fkWorkorderNumber;
 
     @TableField("FK_job_number")
-    private Integer fkJobNumber;
+    private Long fkJobNumber;
 
     private String maintenanceRecord;
