+++ src/main/java/com/example/repair/entity/StudentAccount.java
@@ -30,7 +30,7 @@ public class StudentAccount implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "student_number", type = IdType.INPUT)
-    private Integer studentNumber;
+    private Long studentNumber;
 
     private String name;
